Ansar Rental/backend/ansar_rental_backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-otp==1.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psutil==5.9.4
PyJWT==1.7.1
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
